requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
sqlparse==0.4.3
tabulate==0.9.0
Werkzeug==2.2.2
yoyo-migrations==8.2.0
zipp==3.11.0
